+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 ## [Unreleased]
+### Fixed
+- Start time issue for Virtual parent Items, by @HardNorth
+### Changed
+- Client version updated on [5.3.9](https://github.com/reportportal/client-java/releases/tag/5.3.9), by @HardNorth
 
 ## [5.3.0]
 ### Added
 - Initial release of the Cucumber agent, by @HardNorth
-
-## [0.0.0-ALPHA]